Split the s3 package into two new ones: s3 and s3-store. The former contains logic that uses Spark while the latter does not
