+++ s3-store/src/main/java/geotrellis/store/s3/AmazonS3URI.java
@@ -12,7 +12,7 @@
  * express or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package geotrellis.spark.io.s3;
+package geotrellis.store.s3;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URI;
