+++ s3-store/src/main/scala/geotrellis/store/s3/S3CollectionLayerReader.scala
@@ -14,14 +14,12 @@
  * limitations under the License.
  */
 
-package geotrellis.spark.io.s3
+package geotrellis.store.s3
 
-import geotrellis.layers.{ContextCollection, LayerId}
 import geotrellis.tiling._
-import geotrellis.spark._
-import geotrellis.spark.io._
-import geotrellis.layers.io.avro._
-import geotrellis.layers.io.index._
+import geotrellis.layers._
+import geotrellis.layers.avro._
+import geotrellis.layers.index._
 import geotrellis.util._
 
 import software.amazon.awssdk.services.s3.model._
