+++ s3-store/src/main/scala/geotrellis/store/s3/S3LayerDeleter.scala
@@ -14,15 +14,14 @@
  * limitations under the License.
  */
 
-package geotrellis.spark.io.s3
+package geotrellis.store.s3
 
-import geotrellis.spark.io._
+import geotrellis.layers._
 
 import software.amazon.awssdk.services.s3.model._
 import software.amazon.awssdk.services.s3.S3Client
 
 import com.typesafe.scalalogging.LazyLogging
-import geotrellis.layers.LayerId
 
 import scala.collection.JavaConverters._
