+++ s3-store/src/main/scala/geotrellis/store/s3/cog/S3COGCollectionLayerReader.scala
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-package geotrellis.spark.io.s3.cog
+package geotrellis.store.s3.cog
 
 import geotrellis.raster.CellGrid
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader
 import geotrellis.tiling.{Boundable, SpatialComponent}
-import geotrellis.spark._
-import geotrellis.spark.io._
-import geotrellis.spark.io.cog._
-import geotrellis.layers.io.index._
-import geotrellis.spark.io.s3._
-import geotrellis.spark.io.s3.conf.S3Config
+import geotrellis.layers._
+import geotrellis.layers.cog._
+import geotrellis.layers.index._
+import geotrellis.store.s3._
+import geotrellis.store.s3.conf.S3Config
 import geotrellis.util._
 
 import software.amazon.awssdk.services.s3.S3Client
@@ -33,9 +32,6 @@ import software.amazon.awssdk.services.s3.model._
 import com.typesafe.scalalogging.LazyLogging
 import spray.json.JsonFormat
 import java.net.URI
-
-import geotrellis.layers.io.cog.{COGCollectionLayerReader, ZoomRange}
-import geotrellis.layers.{LayerId, TileLayerMetadata}
 
 import scala.reflect.ClassTag
