+++ s3-store/src/main/scala/geotrellis/store/s3/cog/S3COGValueReader.scala
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
-package geotrellis.spark.io.s3.cog
+package geotrellis.store.s3.cog
 
 import geotrellis.raster._
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader
 import geotrellis.tiling.SpatialComponent
-import geotrellis.spark._
-import geotrellis.spark.io._
-import geotrellis.spark.io.cog._
-import geotrellis.spark.io.s3._
-import geotrellis.layers.io.index._
+import geotrellis.layers._
+import geotrellis.layers.cog._
+import geotrellis.layers.index._
+import geotrellis.store.s3._
 import geotrellis.util._
 
 import software.amazon.awssdk.services.s3.model._
@@ -33,8 +32,6 @@ import spray.json._
 import scala.reflect.ClassTag
 import java.net.URI
 
-import geotrellis.layers.LayerId
-import geotrellis.layers.io.cog.{COGReader, OverzoomingCOGValueReader, ZoomRange}
 
 class S3COGValueReader(
   val attributeStore: AttributeStore,
