+++ s3/src/main/scala/geotrellis/spark/io/s3/Implicits.scala
@@ -1,0 +1,10 @@
+package geotrellis.spark.io.s3
+
+import org.apache.spark.rdd.RDD
+
+
+object Implicits extends Implicits
+
+trait Implicits {
+  implicit class withSaveToS3Methods[K](rdd: RDD[(K, Array[Byte])]) extends SaveToS3Methods(rdd)
+}
