+++ s3/src/main/scala/geotrellis/spark/io/s3/S3GeoTiffInfoReader.scala
@@ -19,13 +19,14 @@ package geotrellis.spark.io.s3
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader.GeoTiffInfo
 import geotrellis.raster.io.geotiff.tags.TiffTags
+import geotrellis.store.s3._
+import geotrellis.store.s3.util.S3RangeReader
 import geotrellis.spark.io._
-import geotrellis.spark.io.s3.util.S3RangeReader
 import geotrellis.util.ByteReader
 
 import software.amazon.awssdk.services.s3.S3Client
 import software.amazon.awssdk.services.s3.model.ListObjectsV2Request
-import geotrellis.spark.io.s3.AmazonS3URI
+
 import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
