+++ s3/src/main/scala/geotrellis/spark/io/s3/S3LayerReader.scala
@@ -17,16 +17,21 @@
 package geotrellis.spark.io.s3
 
 import geotrellis.tiling._
+import geotrellis.layers._
+import geotrellis.layers.avro._
+import geotrellis.layers.index._
+import geotrellis.store.s3._
+import geotrellis.store.s3.conf.S3Config
 import geotrellis.spark._
 import geotrellis.spark.io._
-import geotrellis.spark.io.s3.conf.S3Config
-import geotrellis.layers.io.avro._
-import geotrellis.layers.io.index._
 import geotrellis.util._
+
 import com.typesafe.scalalogging.LazyLogging
-import geotrellis.layers.LayerId
+
 import org.apache.spark.SparkContext
+
 import spray.json.JsonFormat
+
 import software.amazon.awssdk.services.s3.S3Client
 import software.amazon.awssdk.services.s3.model._
