+++ s3/src/main/scala/geotrellis/spark/io/s3/S3LayerWriter.scala
@@ -17,20 +17,25 @@
 package geotrellis.spark.io.s3
 
 import geotrellis.tiling._
+import geotrellis.layers._
+import geotrellis.layers.avro._
+import geotrellis.layers.avro.codecs._
+import geotrellis.layers.index._
+import geotrellis.layers.merge.Mergable
+import geotrellis.store.s3._
+import geotrellis.store.s3.conf.S3Config
 import geotrellis.spark._
 import geotrellis.spark.io._
-import geotrellis.spark.io.s3.conf.S3Config
-import geotrellis.layers.io.avro._
-import geotrellis.layers.io.avro.codecs._
-import geotrellis.layers.io.index._
 import geotrellis.spark.merge._
 import geotrellis.util._
+
 import com.typesafe.scalalogging.LazyLogging
+
 import org.apache.spark.rdd.RDD
+
 import software.amazon.awssdk.services.s3.model.PutObjectRequest
 import software.amazon.awssdk.services.s3.S3Client
-import geotrellis.layers.merge.Mergable
-import geotrellis.layers.{LayerId, Metadata}
+
 import spray.json._
 
 import scala.reflect._
