+++ s3/src/test/scala/geotrellis/spark/io/s3/S3AttributeStoreSpec.scala
@@ -17,6 +17,7 @@
 package geotrellis.spark.io.s3
 
 import geotrellis.layers.LayerId
+import geotrellis.store.s3._
 import geotrellis.spark._
 import geotrellis.spark.io._
 import geotrellis.spark.io.s3.testkit._
