+++ s3/src/test/scala/geotrellis/spark/io/s3/S3SpatialSpec.scala
@@ -16,12 +16,13 @@
 
 package geotrellis.spark.io.s3
 
-import geotrellis.layers.TileLayerMetadata
 import geotrellis.raster.Tile
 import geotrellis.tiling.SpatialKey
+import geotrellis.layers._
+import geotrellis.layers.index._
+import geotrellis.store.s3._
 import geotrellis.spark.io._
 import geotrellis.spark.io.s3.testkit._
-import geotrellis.spark.io.index._
 import geotrellis.spark._
 import geotrellis.spark.testkit.io._
 import geotrellis.spark.testkit.testfiles.TestFiles
