+++ s3/src/test/scala/geotrellis/spark/io/s3/cog/COGS3AttributeStoreSpec.scala
@@ -16,6 +16,8 @@
 
 package geotrellis.spark.io.s3.cog
 
+import geotrellis.layers._
+import geotrellis.store.s3._
 import geotrellis.spark._
 import geotrellis.spark.io._
 import geotrellis.spark.io.cog._
