+++ s3/src/test/scala/geotrellis/spark/io/s3/util/S3RangeReaderSpec.scala
@@ -18,6 +18,7 @@ package geotrellis.spark.io.s3.util
 
 import java.nio.file.{ Paths, Files }
 import java.nio.ByteBuffer
+import geotrellis.store.s3.util._
 import geotrellis.util._
 import geotrellis.spark.io.s3._
 import geotrellis.spark.io.s3.testkit._
